move quote to separate dir
