+++ ydb/core/persqueue/pqtablet/partition/user_info.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "account_read_quoter.h"
+//#include "account_read_quoter.h"
 
 #include "working_time_counter.h"
 #include "subscriber.h"
