+++ ydb/core/persqueue/pqtablet/ya.make
@@ -25,6 +25,7 @@ RECURSE(
     blob
     common
     partition
+    quote
     readproxy
 )
